Title: 특정 구간의 원소 평균값, Time: 126ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240331/특정 구간의 원소 평균값/elemental-mean-value-for-a-particular-interval.java
@@ -19,7 +19,7 @@ public class Main {
                 avg = avg / (j - i + 1);
                 
                 for(int k = i; k <= j; k++){
-                    if(avg == arr[k]){
+                    if((int)avg == arr[k]){
                         cnt++;
                         break;
                     }
